refactor: break dependency_checker_service.py into manageable modules
Extract validation, caching, and installation guide functionality into
separate helper modules for better code organization:
- dependency_validators.py: DependencyValidator and ProviderValidator classes
- cache_manager.py: CacheHelper class for cache operations
- installation_guides.py: InstallationGuideHelper for installation guides

Main service reduced from 741 to 391 lines while maintaining backward
compatibility. Updated test patches to use internal validator objects.
